perf: protect group_leader from races that cause ctx double-free
When moving a group_leader perf event from a software-context to a hardware-context, there's a race in checking and updating that context.

The existing locking solution doesn't work; note that it tries to grab a lock inside the group_leader's context object, which you can only get at by going through a pointer that should be protected from these races.

To avoid that problem, and to produce a simple solution, we can just use a lock per group_leader to protect all checks on the group_leader's context.

The new lock is grabbed and released when no context locks are held.